Changes 04.05.2016
- Changed column sizes, appearance, color, functionality in the tables
- Added contact Director
- Added an administrator to contact the Director, and help
- Added link to any set of meetings in the news and in selecting the
place and date of the meeting
- Added a link to all completed meetings in the News
- Changed the appearance of news on two separate columns
- Fixed a bug (do not know if such existed already in the league) If you
changed your name or surname
the site being shown a new name in the schedule, and that name did not
update everywhere
the site for which was not available for that user various options, for
example. accepting match
- Added bookmark matches for the administrator (for now only the view,
in the next version of functionality: (
  - Matches approved by one person, the possibility of approval instead
of the other person.
  - Matches not approved by any person, the ability to select the winner
of the match and assigned in sets 4: 0
)